+++ src/main/java/org/bytedeco/javacpp/CharPointer.java
@@ -76,6 +76,9 @@ public class CharPointer extends Pointer {
     public CharPointer(long size) {
         try {
             allocateArray(size);
+            if (size > 0 && address == 0) {
+                throw new OutOfMemoryError("Cannot allocate new CharPointer(" + size + ")");
+            }
         } catch (UnsatisfiedLinkError e) {
             throw new RuntimeException("No native JavaCPP library in memory. (Has Loader.load() been called?)", e);
         }